+++ src/main/java/Main.java
@@ -2,7 +2,6 @@
 import dao.LukuvinkkiDao;
 import io.ConsoleIO;
 import ui.ConsoleUI;
-import ui.ConsoleUIbeta;
 import java.io.IOException;
 
 public class Main {
@@ -12,7 +11,6 @@ public class Main {
         LukuvinkkiDao dao = new LukuvinkkiDao();
 
         ConsoleUI ui = new ConsoleUI(console, dao);
-//        ConsoleUI ui = new ConsoleUIbeta(console, dao);
         ui.run();
     }
